🔧 Add : 보드 삭제 서비스 메서드 수정
🔧 Add : 보드 삭제 서비스 메서드 수정
- 이미지 null 체크 후 삭제 진행하는 코드 추가

+++ src/main/java/com/currency/teamflow/domain/board/service/BoardService.java
@@ -116,7 +116,12 @@ public class BoardService {
 	public void deleteBoard(Long boardId) {
 		//보드 정보 가져오기
 		Board board = boardRepository.findByIdOrElseThrow(boardId);
-		attachmentService.deleteAttachmentForBoard(board.getImageUrl());
+
+		//이미지가 저장이 되어있다면 삭제
+		if(board.getImageUrl() != null && !board.getImageUrl().isEmpty()){
+			//이미지 정보 삭제하기
+			attachmentService.deleteAttachmentForBoard(board.getImageUrl());
+		}
 		boardRepository.delete(board);
 	}
